+++ mocks/application/mock_UpdateUserPhoneUseCase.go
@@ -23,7 +23,7 @@ func (_m *MockUpdateUserPhoneUseCase) EXPECT() *MockUpdateUserPhoneUseCase_Expec
 }
 
 // Execute provides a mock function with given fields: ctx, id, phone
-func (_m *MockUpdateUserPhoneUseCase) Execute(ctx context.Context, id int, phone domain.Phone) error {
+func (_m *MockUpdateUserPhoneUseCase) Execute(ctx context.Context, id int64, phone domain.Phone) error {
 	ret := _m.Called(ctx, id, phone)
 
 	if len(ret) == 0 {
@@ -31,7 +31,7 @@ func (_m *MockUpdateUserPhoneUseCase) Execute(ctx context.Context, id int, phone
 	}
 
 	var r0 error
-	if rf, ok := ret.Get(0).(func(context.Context, int, domain.Phone) error); ok {
+	if rf, ok := ret.Get(0).(func(context.Context, int64, domain.Phone) error); ok {
 		r0 = rf(ctx, id, phone)
 	} else {
 		r0 = ret.Error(0)
@@ -47,15 +47,15 @@ type MockUpdateUserPhoneUseCase_Execute_Call struct {
 
 // Execute is a helper method to define mock.On call
 //   - ctx context.Context
-//   - id int
+//   - id int64
 //   - phone domain.Phone
 func (_e *MockUpdateUserPhoneUseCase_Expecter) Execute(ctx interface{}, id interface{}, phone interface{}) *MockUpdateUserPhoneUseCase_Execute_Call {
 	return &MockUpdateUserPhoneUseCase_Execute_Call{Call: _e.mock.On("Execute", ctx, id, phone)}
 }
 
-func (_c *MockUpdateUserPhoneUseCase_Execute_Call) Run(run func(ctx context.Context, id int, phone domain.Phone)) *MockUpdateUserPhoneUseCase_Execute_Call {
+func (_c *MockUpdateUserPhoneUseCase_Execute_Call) Run(run func(ctx context.Context, id int64, phone domain.Phone)) *MockUpdateUserPhoneUseCase_Execute_Call {
 	_c.Call.Run(func(args mock.Arguments) {
-		run(args[0].(context.Context), args[1].(int), args[2].(domain.Phone))
+		run(args[0].(context.Context), args[1].(int64), args[2].(domain.Phone))
 	})
 	return _c
 }
@@ -65,7 +65,7 @@ func (_c *MockUpdateUserPhoneUseCase_Execute_Call) Return(_a0 error) *MockUpdate
 	return _c
 }
 
-func (_c *MockUpdateUserPhoneUseCase_Execute_Call) RunAndReturn(run func(context.Context, int, domain.Phone) error) *MockUpdateUserPhoneUseCase_Execute_Call {
+func (_c *MockUpdateUserPhoneUseCase_Execute_Call) RunAndReturn(run func(context.Context, int64, domain.Phone) error) *MockUpdateUserPhoneUseCase_Execute_Call {
 	_c.Call.Return(run)
 	return _c
 }
